requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
dnspython>=2.7.0
python-whois>=0.9.5
httpx>=0.28.0
